+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/PanZoomController.java
@@ -98,10 +98,10 @@ public class PanZoomController {
         }
 
         @WrapForJNI(calledFrom = "ui")
-        private native @InputResult int handleMotionEvent(
+        private native void handleMotionEvent(
                int action, int actionIndex, long time, int metaState,  float screenX, float screenY,
                int pointerId[], float x[], float y[], float orientation[], float pressure[],
-               float toolMajor[], float toolMinor[]);
+               float toolMajor[], float toolMinor[], GeckoResult<Integer> result);
 
         @WrapForJNI(calledFrom = "ui")
         private native @InputResult int handleScrollEvent(
@@ -150,10 +150,17 @@ public class PanZoomController {
 
     /* package */ final NativeProvider mNative = new NativeProvider();
 
-    private @InputResult int handleMotionEvent(final MotionEvent event) {
+    private void handleMotionEvent(final MotionEvent event) {
+        handleMotionEvent(event, null);
+    }
+
+    private void handleMotionEvent(final MotionEvent event, final GeckoResult<Integer> result) {
         if (!mAttached) {
             mQueuedEvents.add(new Pair<>(EVENT_SOURCE_MOTION, event));
-            return INPUT_RESULT_HANDLED;
+            if (result != null) {
+                result.complete(INPUT_RESULT_HANDLED);
+            }
+            return;
         }
 
         final int action = event.getActionMasked();
@@ -162,7 +169,10 @@ public class PanZoomController {
         if (action == MotionEvent.ACTION_DOWN) {
             mLastDownTime = event.getDownTime();
         } else if (mLastDownTime != event.getDownTime()) {
-            return INPUT_RESULT_UNHANDLED;
+            if (result != null) {
+                result.complete(INPUT_RESULT_UNHANDLED);
+            }
+            return;
         }
 
         final int[] pointerId = new int[count];
@@ -200,9 +210,9 @@ public class PanZoomController {
             mSession.onScreenOriginChanged((int)screenX, (int)screenY);
         }
 
-        return mNative.handleMotionEvent(action, event.getActionIndex(), event.getEventTime(),
-                                         event.getMetaState(), screenX, screenY, pointerId, x, y,
-                                         orientation, pressure, toolMajor, toolMinor);
+        mNative.handleMotionEvent(action, event.getActionIndex(), event.getEventTime(),
+                                  event.getMetaState(), screenX, screenY, pointerId, x, y,
+                                  orientation, pressure, toolMajor, toolMinor, result);
     }
 
     private @InputResult int handleScrollEvent(final MotionEvent event) {
@@ -314,15 +324,42 @@ public class PanZoomController {
      * display surface.
      *
      * @param event MotionEvent to process.
-     * @return One of the {@link PanZoomController#INPUT_RESULT_UNHANDLED INPUT_RESULT_*}) constants indicating how the event was handled.
      */
-    public @InputResult int onTouchEvent(final @NonNull MotionEvent event) {
+    public void onTouchEvent(final @NonNull MotionEvent event) {
         ThreadUtils.assertOnUiThread();
 
         if (!sTreatMouseAsTouch && event.getToolType(0) == MotionEvent.TOOL_TYPE_MOUSE) {
-            return handleMouseEvent(event);
+            handleMouseEvent(event);
+            return;
         }
-        return handleMotionEvent(event);
+        handleMotionEvent(event);
+    }
+
+    /**
+     * Process a touch event through the pan-zoom controller. Treat any mouse events as
+     * "touch" rather than as "mouse". Pointer coordinates should be relative to the
+     * display surface.
+     *
+     * NOTE: It is highly recommended to only call this with ACTION_DOWN or in otherwise
+     * limited capacity. Returning a GeckoResult for every touch event will generate
+     * a lot of allocations and unnecessary GC pressure. Instead, prefer to call
+     * {@link #onTouchEvent(MotionEvent)}.
+     *
+     * @param event MotionEvent to process.
+     * @return A GeckoResult resolving to one of the
+     *         {@link PanZoomController#INPUT_RESULT_UNHANDLED INPUT_RESULT_*}) constants indicating
+     *         how the event was handled.
+     */
+    public @NonNull GeckoResult<Integer> onTouchEventForResult(final @NonNull MotionEvent event) {
+        ThreadUtils.assertOnUiThread();
+
+        if (!sTreatMouseAsTouch && event.getToolType(0) == MotionEvent.TOOL_TYPE_MOUSE) {
+            return GeckoResult.fromValue(handleMouseEvent(event));
+        }
+
+        final GeckoResult<Integer> result = new GeckoResult<>();
+        handleMotionEvent(event, result);
+        return result;
     }
 
     /**
@@ -331,15 +368,14 @@ public class PanZoomController {
      * display surface.
      *
      * @param event MotionEvent to process.
-     * @return One of the {@link PanZoomController#INPUT_RESULT_UNHANDLED INPUT_RESULT_*}) constants indicating how the event was handled.
      */
-    public @InputResult int onMouseEvent(final @NonNull MotionEvent event) {
+    public void onMouseEvent(final @NonNull MotionEvent event) {
         ThreadUtils.assertOnUiThread();
 
         if (event.getToolType(0) == MotionEvent.TOOL_TYPE_MOUSE) {
-            return handleMouseEvent(event);
+            return;
         }
-        return handleMotionEvent(event);
+        handleMotionEvent(event);
     }
 
     @Override
@@ -353,9 +389,8 @@ public class PanZoomController {
      * display surface.
      *
      * @param event MotionEvent to process.
-     * @return One of the {@link PanZoomController#INPUT_RESULT_UNHANDLED INPUT_RESULT_*}) indicating how the event was handled.
      */
-    public @InputResult int onMotionEvent(final @NonNull MotionEvent event) {
+    public void onMotionEvent(final @NonNull MotionEvent event) {
         ThreadUtils.assertOnUiThread();
 
         final int action = event.getActionMasked();
@@ -365,15 +400,13 @@ public class PanZoomController {
             } else if ((InputDevice.getDevice(event.getDeviceId()) != null) &&
                        (InputDevice.getDevice(event.getDeviceId()).getSources() &
                         InputDevice.SOURCE_TOUCHPAD) == InputDevice.SOURCE_TOUCHPAD) {
-                return INPUT_RESULT_UNHANDLED;
+                return;
             }
-            return handleScrollEvent(event);
+            handleScrollEvent(event);
         } else if ((action == MotionEvent.ACTION_HOVER_MOVE) ||
                    (action == MotionEvent.ACTION_HOVER_ENTER) ||
                    (action == MotionEvent.ACTION_HOVER_EXIT)) {
-            return handleMouseEvent(event);
-        } else {
-            return INPUT_RESULT_UNHANDLED;
+            handleMouseEvent(event);
         }
     }